JSLINT + Bug Fix
There is a bug where if you type a title that gets converted to a slug
(using default settings), and then modify the slug and then tab out. if
you're using default settings all your dashes will be removed. I added
to the map "2d":"-" which will replace dashes with dashes which works
well for if your editing the title and also if you're editing the slug